Update Dashboard.vue
Added negative bottom margin to Alerts section to cancel out the bottom margin added by each alert. (This will be refactored in the future when the alerts are all part of the same expansion panels element.)

This is so the whitespace between the Alerts section and the To Do section is the same as between the To Do section and the Recent Filing History section.